+++ CMakeLists.txt
@@ -515,6 +515,7 @@ endif()
 
 if (NOT ANDROID AND NOT WEBGL AND NOT IOS)
     add_subdirectory(${LIBRARIES}/bluegl)
+    add_subdirectory(${LIBRARIES}/filamentapp)
     add_subdirectory(${LIBRARIES}/filagui)
     add_subdirectory(${LIBRARIES}/imageio)
     add_subdirectory(${LIBRARIES}/matdbg)
