+++ libs/filamentapp/materials/depthVisualizer.mat
@@ -1,0 +1,14 @@
+material {
+    name : depthVisualizer,
+    shadingModel : unlit,
+}
+
+fragment {
+    void material(inout MaterialInputs material) {
+        prepareMaterial(material);
+        const float n = 0.1;
+        const float f = 35.0;
+        float linearZ = (2.0 * n) / (f + n - gl_FragCoord.z * (f - n));
+        material.baseColor.rgb = float3(linearZ);
+    }
+}
