--- samples/lucy_utils.h
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef TNT_FILAMENT_SAMPLES_LUCY_UTILS_H
-#define TNT_FILAMENT_SAMPLES_LUCY_UTILS_H
-
-#include <filament/Engine.h>
-#include <filament/Texture.h>
-#include <filament/Box.h>
-
-#include <math/mat4.h>
-
-#include "generated/resources/gltf_lucy.h"
-
-#define RESOURCE_ARGS(name) GLTF_LUCY_##name##_DATA, GLTF_LUCY_##name##_SIZE
-
-namespace LucyUtils {
-    enum ImageOp { BLIT, HBLUR, VBLUR, MIX };
-    utils::Entity createQuad(filament::Engine* engine, filament::Texture* tex, ImageOp op,
-            filament::Texture* secondary = nullptr);
-    utils::Entity createDisk(filament::Engine* engine, filament::Texture* reflection);
-    filament::math::mat4f fitIntoUnitCube(const filament::Aabb& bounds);
-}
-
-#endif // #define TNT_FILAMENT_SAMPLES_LUCY_UTILS_H
