+++ samples/strobecolor.cpp
@@ -18,8 +18,8 @@
 #include <filament/View.h>
 #include <filament/Scene.h>
 
-#include "../samples/app/Config.h"
-#include "../samples/app/FilamentApp.h"
+#include "filamentapp/Config.h"
+#include "filamentapp/FilamentApp.h"
 
 #include <cmath>
