+++ samples/texturedquad.cpp
@@ -30,8 +30,8 @@
 
 #include <utils/Path.h>
 
-#include "../samples/app/Config.h"
-#include "../samples/app/FilamentApp.h"
+#include "filamentapp/Config.h"
+#include "filamentapp/FilamentApp.h"
 
 #include <stb_image.h>
